requirements.txt
streamlit==1.41.0
sentence-transformers==3.3.1
faiss-cpu>=1.7.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv==1.0.1
